requirements.txt
arrow==0.6.0
Django==1.8
django-filter==0.11.0
djangorestframework==3.2.3
python-dateutil==2.4.2
six==1.9.0
wheel==0.24.0
